backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
google-cloud-firestore==2.14.0
google-auth==2.28.1
google-api-python-client==2.118.0
google-cloud-aiplatform==1.71.1
google-cloud-storage==2.14.0
vertexai==1.71.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-lazy-fixture==0.6.3
httpx==0.26.0
firebase-admin==6.4.0
debugpy==1.8.0
Pillow==10.2.0
replicate==1.0.7

